# Conflicts:
#	app/src/main/java/com/example/university/View/Main/Screens/FutureTestsView.kt

+++ app/src/main/java/com/example/university/ViewModel/States/PickQuantityUiState.kt
@@ -5,7 +5,7 @@ import com.example.university.theme.PHColors
 
 data class PickQuantityUiState(
     val wordsQuantity: Int = 1,
-    val pickedQuantity: Int = 1,
+    val pickedQuantity: Float = 1f,
     val colorScheme: Colors = PHColors,
     val isRememberPresent: Boolean = false,
     val pickedWords: String = ""
